requirements.txt
psycopg2==2.4.4
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.6
django-heroku==0.0.0
gunicorn==20.1.0
markdown2==2.4.3
Pillow==9.2.0
python-decouple==3.6
sqlparse==0.4.2
whitenoise==6.2.0
